requirements.txt
yfinance
pandas
numpy
ta
python-dotenv
requests
schedule
SQLAlchemy
